backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfminer.six==20231228

numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
joblib==1.3.2
threadpoolctl==3.5.0

regex==2024.5.15
nltk==3.8.1

# novo
httpx==0.28.1
